+++ docs/examples_literate/general_examples/optimal_advertising.jl
@@ -46,7 +46,7 @@ D = Variable(m, n);
 Si = [min(R[i]*dot(P[i,:], D[i,:]'), B[i]) for i=1:m];
 problem = maximize(sum(Si),
                [D >= 0, sum(D, dims=1)' <= T, sum(D, dims=2) >= c]);
-solve!(problem, SCS.Optimizer(verbose=0));
+solve!(problem, () -> SCS.Optimizer(verbose=0));
 
 #-