+++ docs/examples_literate/optimization_with_complex_variables/Fidelity in Quantum Information Theory.jl
@@ -33,7 +33,7 @@ Z = ComplexVariable(n,n)
 objective = 0.5*real(tr(Z+Z'))
 constraint = [P Z;Z' Q] ⪰ 0
 problem = maximize(objective,constraint)
-solve!(problem, SCS.Optimizer(verbose=0))
+solve!(problem, () -> SCS.Optimizer(verbose=0))
 computed_fidelity = evaluate(objective)
 
 #-